moved visuals to src/saev/scripts/
